[Experiment] C++ interface to create a lowered backend module.
[ghstack-poisoned]
