+++ torch/csrc/jit/backends/backend_detail.h
@@ -46,6 +46,11 @@ std::function<void(Stack&)> getExecuteFunc() {
     push(stack, res);
   };
 }
+
+Module codegen_backend_module(const std::string& backend_name,
+                              const Module& orig_module,
+                              const c10::Dict<IValue, IValue>& method_compile_spec,
+                              c10::DictTypePtr any_dict_ty);
 } // namespace detail
 } // namespace jit
 } // namespace torch
